update about info on README …
I’ll still write a blog post about all this, with a link to the first
documented occurrence of my writing style shift. This should be the
second, and I’m thinking about it right now. periods are as good as
commas to give text a reading flow stop, albeit a bigger one. and
usually there comes an upper case letter to emphasise that, maybe
making reading and spotting the small dot easier. still I rather,
instead, use it to denote a higher voice tone. why Not? :)